Stable, acceptable version

+++ physics.h
@@ -3,15 +3,15 @@ namespace R_physics
 	extern float groundHeight;
 	extern float playerWidth;
 	extern float playerHeight;
-	extern int dotWidth;
-	extern int dotHeight;
-	extern int dashWidth;
-	extern int dashHeight;
+	extern float dotWidth;
+	extern float dotHeight;
+	extern float dashWidth;
+	extern float dashHeight;
 	extern float playerX, playerY;		
 	extern char curLetter;				//The current letter to be displayed on the screen
 	extern bool jumpForceOn;			//Set to true to make player jump and hover. Set to false to make player fall if in the air
 	extern long long SCORE;
-	extern int triPos[60][2];
+	extern float triPos[60][2];
 
 	float getPlayerX();
     float getPlayerY();					//x and y coordinates of player in physics system. SCALE BEFORE USE.
